Replace PyObject with type alias for Py<T>
