+++ src/instance.rs
@@ -1,9 +1,10 @@
 // Copyright (c) 2017-present PyO3 Project and Contributors
-use crate::err::{PyErr, PyResult};
+use crate::conversion::{PyTryFrom, ToBorrowedObject};
+use crate::err::{PyDowncastError, PyErr, PyResult};
 use crate::gil;
-use crate::object::PyObject;
 use crate::pycell::{PyBorrowError, PyBorrowMutError, PyCell};
 use crate::type_object::PyBorrowFlagLayout;
+use crate::types::{PyDict, PyTuple};
 use crate::{
     ffi, AsPyPointer, FromPyObject, IntoPy, IntoPyPointer, PyAny, PyClass, PyClassInitializer,
     PyRef, PyRefMut, PyTypeInfo, Python, ToPyObject,
@@ -114,6 +115,11 @@ where
 }
 
 impl<T> Py<T> {
+    /// For internal conversions
+    pub(crate) unsafe fn from_not_null(ptr: NonNull<ffi::PyObject>) -> Self {
+        Self(ptr, PhantomData)
+    }
+
     /// Creates a `Py<T>` instance for the given FFI pointer.
     ///
     /// This moves ownership over the pointer into the `Py<T>`.
@@ -186,6 +192,168 @@ impl<T> Py<T> {
         let pointer = self.0;
         mem::forget(self);
         pointer
+    }
+
+    /// Constructs a `PyObject` from the result of a Python FFI call that
+    /// returns a new reference (owned pointer).
+    /// Returns `None` if the pointer is NULL.
+    pub unsafe fn from_owned_ptr_or_opt(_py: Python, ptr: *mut ffi::PyObject) -> Option<Self> {
+        match NonNull::new(ptr) {
+            Some(nonnull_ptr) => Some(Py(nonnull_ptr, PhantomData)),
+            None => None,
+        }
+    }
+
+    /// Creates a `PyObject` instance for the given Python FFI pointer.
+    /// Calls `Py_INCREF()` on the ptr.
+    /// Returns `Err(PyErr)` if the pointer is NULL.
+    pub unsafe fn from_borrowed_ptr_or_err(py: Python, ptr: *mut ffi::PyObject) -> PyResult<Self> {
+        if ptr.is_null() {
+            Err(PyErr::fetch(py))
+        } else {
+            Ok(Self::from_borrowed_ptr(py, ptr))
+        }
+    }
+
+    /// Creates a `PyObject` instance for the given Python FFI pointer.
+    /// Calls `Py_INCREF()` on the ptr.
+    /// Returns `None` if the pointer is NULL.
+    pub unsafe fn from_borrowed_ptr_or_opt(py: Python, ptr: *mut ffi::PyObject) -> Option<Self> {
+        if ptr.is_null() {
+            None
+        } else {
+            Some(Self::from_borrowed_ptr(py, ptr))
+        }
+    }
+
+    /// Returns whether the object is considered to be None.
+    ///
+    /// This is equivalent to the Python expression `self is None`.
+    pub fn is_none(&self, _py: Python) -> bool {
+        unsafe { ffi::Py_None() == self.as_ptr() }
+    }
+
+    /// Returns whether the object is considered to be true.
+    ///
+    /// This is equivalent to the Python expression `bool(self)`.
+    pub fn is_true(&self, py: Python) -> PyResult<bool> {
+        let v = unsafe { ffi::PyObject_IsTrue(self.as_ptr()) };
+        if v == -1 {
+            Err(PyErr::fetch(py))
+        } else {
+            Ok(v != 0)
+        }
+    }
+
+    /// Casts the PyObject to a concrete Python object type.
+    ///
+    /// This can cast only to native Python types, not types implemented in Rust.
+    pub fn cast_as<'p, D>(&'p self, py: Python<'p>) -> Result<&'p D, PyDowncastError>
+    where
+        D: PyTryFrom<'p>,
+    {
+        D::try_from(unsafe { py.from_borrowed_ptr::<PyAny>(self.as_ptr()) })
+    }
+
+    /// Extracts some type from the Python object.
+    ///
+    /// This is a wrapper function around `FromPyObject::extract()`.
+    pub fn extract<'p, D>(&'p self, py: Python<'p>) -> PyResult<D>
+    where
+        D: FromPyObject<'p>,
+    {
+        FromPyObject::extract(unsafe { py.from_borrowed_ptr(self.as_ptr()) })
+    }
+
+    /// Retrieves an attribute value.
+    ///
+    /// This is equivalent to the Python expression `self.attr_name`.
+    pub fn getattr<N>(&self, py: Python, attr_name: N) -> PyResult<PyObject>
+    where
+        N: ToPyObject,
+    {
+        attr_name.with_borrowed_ptr(py, |attr_name| unsafe {
+            PyObject::from_owned_ptr_or_err(py, ffi::PyObject_GetAttr(self.as_ptr(), attr_name))
+        })
+    }
+
+    /// Calls the object.
+    ///
+    /// This is equivalent to the Python expression `self(*args, **kwargs)`.
+    pub fn call(
+        &self,
+        py: Python,
+        args: impl IntoPy<Py<PyTuple>>,
+        kwargs: Option<&PyDict>,
+    ) -> PyResult<PyObject> {
+        let args = args.into_py(py).into_ptr();
+        let kwargs = kwargs.into_ptr();
+        let result = unsafe {
+            PyObject::from_owned_ptr_or_err(py, ffi::PyObject_Call(self.as_ptr(), args, kwargs))
+        };
+        unsafe {
+            ffi::Py_XDECREF(args);
+            ffi::Py_XDECREF(kwargs);
+        }
+        result
+    }
+
+    /// Calls the object with only positional arguments.
+    ///
+    /// This is equivalent to the Python expression `self(*args)`.
+    pub fn call1(&self, py: Python, args: impl IntoPy<Py<PyTuple>>) -> PyResult<PyObject> {
+        self.call(py, args, None)
+    }
+
+    /// Calls the object without arguments.
+    ///
+    /// This is equivalent to the Python expression `self()`.
+    pub fn call0(&self, py: Python) -> PyResult<PyObject> {
+        self.call(py, (), None)
+    }
+
+    /// Calls a method on the object.
+    ///
+    /// This is equivalent to the Python expression `self.name(*args, **kwargs)`.
+    pub fn call_method(
+        &self,
+        py: Python,
+        name: &str,
+        args: impl IntoPy<Py<PyTuple>>,
+        kwargs: Option<&PyDict>,
+    ) -> PyResult<PyObject> {
+        name.with_borrowed_ptr(py, |name| unsafe {
+            let args = args.into_py(py).into_ptr();
+            let kwargs = kwargs.into_ptr();
+            let ptr = ffi::PyObject_GetAttr(self.as_ptr(), name);
+            if ptr.is_null() {
+                return Err(PyErr::fetch(py));
+            }
+            let result = PyObject::from_owned_ptr_or_err(py, ffi::PyObject_Call(ptr, args, kwargs));
+            ffi::Py_DECREF(ptr);
+            ffi::Py_XDECREF(args);
+            ffi::Py_XDECREF(kwargs);
+            result
+        })
+    }
+
+    /// Calls a method on the object with only positional arguments.
+    ///
+    /// This is equivalent to the Python expression `self.name(*args)`.
+    pub fn call_method1(
+        &self,
+        py: Python,
+        name: &str,
+        args: impl IntoPy<Py<PyTuple>>,
+    ) -> PyResult<PyObject> {
+        self.call_method(py, name, args, None)
+    }
+
+    /// Calls a method on the object with no arguments.
+    ///
+    /// This is equivalent to the Python expression `self.name()`.
+    pub fn call_method0(&self, py: Python, name: &str) -> PyResult<PyObject> {
+        self.call_method(py, name, (), None)
     }
 }
 
@@ -274,13 +442,22 @@ impl<T> IntoPyPointer for Py<T> {
     }
 }
 
-// Native types `&T` can be converted to `Py<T>`
-impl<'a, T> std::convert::From<&'a T> for Py<T>
+impl<T> std::convert::From<&'_ T> for PyObject
 where
     T: AsPyPointer + PyNativeType,
 {
-    fn from(obj: &'a T) -> Self {
+    fn from(obj: &T) -> Self {
         unsafe { Py::from_borrowed_ptr(obj.py(), obj.as_ptr()) }
+    }
+}
+
+impl<T> std::convert::From<Py<T>> for PyObject
+where
+    T: AsRef<PyAny>,
+{
+    fn from(other: Py<T>) -> Self {
+        let Py(ptr, _) = other;
+        Py(ptr, PhantomData)
     }
 }
 
@@ -337,31 +514,6 @@ impl<T> Drop for Py<T> {
     }
 }
 
-impl<T> std::convert::From<Py<T>> for PyObject {
-    #[inline]
-    fn from(ob: Py<T>) -> Self {
-        unsafe { PyObject::from_not_null(ob.into_non_null()) }
-    }
-}
-
-impl<'a, T> std::convert::From<&'a T> for PyObject
-where
-    T: AsPyPointer + PyNativeType,
-{
-    fn from(ob: &'a T) -> Self {
-        unsafe { Py::<T>::from_borrowed_ptr(ob.py(), ob.as_ptr()) }.into()
-    }
-}
-
-impl<'a, T> std::convert::From<&'a mut T> for PyObject
-where
-    T: AsPyPointer + PyNativeType,
-{
-    fn from(ob: &'a mut T) -> Self {
-        unsafe { Py::<T>::from_borrowed_ptr(ob.py(), ob.as_ptr()) }.into()
-    }
-}
-
 impl<'a, T> FromPyObject<'a> for Py<T>
 where
     T: PyTypeInfo,
@@ -406,16 +558,30 @@ impl<T> std::fmt::Debug for Py<T> {
     }
 }
 
+pub type PyObject = Py<PyAny>;
+
 #[cfg(test)]
 mod test {
-    use super::Py;
-    use crate::ffi;
+    use super::{Py, PyObject};
     use crate::types::PyDict;
-    use crate::{AsPyPointer, Python};
+    use crate::{ffi, AsPyPointer, Python};
+
+    #[test]
+    fn test_call_for_non_existing_method() {
+        let gil = Python::acquire_gil();
+        let py = gil.python();
+        let obj: PyObject = PyDict::new(py).into();
+        assert!(obj.call_method0(py, "asdf").is_err());
+        assert!(obj
+            .call_method(py, "nonexistent_method", (1,), None)
+            .is_err());
+        assert!(obj.call_method0(py, "nonexistent_method").is_err());
+        assert!(obj.call_method1(py, "nonexistent_method", (1,)).is_err());
+    }
 
     #[test]
     fn py_from_dict() {
-        let dict = {
+        let dict: Py<PyDict> = {
             let gil = Python::acquire_gil();
             let py = gil.python();
             let native = PyDict::new(py);
