+++ src/prelude.rs
@@ -12,8 +12,8 @@
 
 pub use crate::err::{PyErr, PyResult};
 pub use crate::gil::GILGuard;
+pub use crate::instance::PyObject;
 pub use crate::instance::{AsPyRef, Py};
-pub use crate::object::PyObject;
 pub use crate::pycell::{PyCell, PyRef, PyRefMut};
 pub use crate::pyclass_init::PyClassInitializer;
 pub use crate::python::Python;
